+++ container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/name_constants.go
@@ -2,4 +2,13 @@ package object_attributes_provider
 
 const (
 	objectNameElementSeparator = "--"
+
+	podNameSuffix             = "pod"
+	serviceNameSuffix         = "service"
+	namespaceSuffix           = "namespace"
+	serviceAccountSuffix      = "service-account"
+	clusterRoleSuffix         = "cluster-role"
+	clusterRoleBindingsSuffix = "cluster-role-bindings"
+	roleSuffix                = "role"
+	roleBindingsSuffix        = "role-bindings"
 )